Move custom controller results to new namespace
