+++ tests/Tgstation.Server.Host.Tests/Swarm/SwarmRpcMapper.cs
@@ -24,6 +24,7 @@ using Newtonsoft.Json;
 using Tgstation.Server.Common.Http;
 using Tgstation.Server.Host.Configuration;
 using Tgstation.Server.Host.Controllers;
+using Tgstation.Server.Host.Controllers.Results;
 using Tgstation.Server.Host.Transfer;
 
 namespace Tgstation.Server.Host.Swarm.Tests
